pyproject.toml
[project]
name = "tds-proj"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bs4>=0.0.2",
    "chromadb>=0.6.3",
    "fastapi>=0.115.12",
    "jsonlines>=4.0.0",
    "langchain>=0.3.25",
    "langchain-openai>=0.3.22",
    "openai>=1.86.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.4",
    "sentence-transformers>=4.1.0",
    "uvicorn>=0.34.3",
]
